Deal with merge commits properly
If a merge commit changes affecting a file, check parents of that
commit for changes, to see if the changes came from another commit.
Note: if both branches affect the file, then we say the merge commit is
the changing commit. We don't look get the actual source of the merged
changes. (ie: make sure that the changes were merged cleanly, then check
for the source of those changes)